chore: more docs and bubble up more API to shutdown
